Module eviov_proto::cs_obs

The observer channel

Structs

AccelEvent
Data for an acceleration event.

ClientHandlerWrapper
Implementation of eviov_proto::Handler for the Client endpoint.

Close
Acknowledges thst the channnel is closed, especially if the client is denied from observation.

Event
Reports a change of object states in the system.

HandshakeRequest
Identifies the client to acknowledge that it is authorized to observe

HandshakeResponse
Response type for HandshakeResponse

Proto
Client-system observer channel

ServerHandlerWrapper
Implementation of eviov_proto::Handler for the Server endpoint.

Enums

Client
An enum of packets sent from the Client endpoint.

EventContent
Contents of an event.

Server
An enum of packets sent from the Server endpoint.

Traits

ClientHandler
Message-specific handler methods for the Client endpoint

ServerHandler
Message-specific handler methods for the Server endpoint