Module eviov_proto::intra

Communication between systems.

Structs

AllowObserveRequest
Allows a session to observing this system.

AllowObserveResponse
Response type for AllowObserveResponse

ClientHandlerWrapper
Implementation of eviov_proto::Handler for the Client endpoint.

FullObject
The full information about an object

Handshake
Binds the connection to a child.

Proto
Communication between systems.

RevokeObserve
Denies a session from observing this system.

ServerHandlerWrapper
Implementation of eviov_proto::Handler for the Server endpoint.

TransferChildRequest
Transfers a child to the peer system.

TransferChildResponse
Response type for TransferChildResponse

TransferYou
Notifies that the child is to be transferred to a new system.

WhereMeRequest
Requests the ECI position and velocity of the child.

WhereMeResponse
Response type for WhereMeResponse

Enums

Client
An enum of packets sent from the Client endpoint.

Server
An enum of packets sent from the Server endpoint.

Traits

ClientHandler
Message-specific handler methods for the Client endpoint

ServerHandler
Message-specific handler methods for the Server endpoint