Module eviov_proto::time

Time synchronization

Structs

ClientHandlerWrapper
Implementation of eviov_proto::Handler for the Client endpoint.

Proto
Time synchronization.

ServerHandlerWrapper
Implementation of eviov_proto::Handler for the Server endpoint.

WhenRequest
Queries the current game time

WhenResponse
Response type for WhenResponse

Enums

Client
An enum of packets sent from the Client endpoint.

Server
An enum of packets sent from the Server endpoint.

Constants

_IMPL_DESERIALIZE_FOR_Client

_IMPL_DESERIALIZE_FOR_Server

_IMPL_DESERIALIZE_FOR_WhenRequest

_IMPL_DESERIALIZE_FOR_WhenResponse

_IMPL_SERIALIZE_FOR_Client

_IMPL_SERIALIZE_FOR_Server

_IMPL_SERIALIZE_FOR_WhenRequest

_IMPL_SERIALIZE_FOR_WhenResponse

Traits

ClientHandler
Message-specific handler methods for the Client endpoint

ServerHandler
Message-specific handler methods for the Server endpoint