Trait eviov_proto::Handler

pub trait Handler: Sized + Send + Sync + 'static {
    type Endpoint: Endpoint;
    fn handle_message<'life0, 'async_trait>(
        &'life0 self, 
        e: <Self::Endpoint as Endpoint>::Peer
    ) -> Pin<Box<dyn Future<Output = Option<Self::Endpoint>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

A generic handler trait for protocols.

Implementations of protocol endpoints should implement the MessageHandler trait generated by the codegen crate instead.

Associated Types

type Endpoint: Endpoint

Loading content...
Required methods

fn handle_message<'life0, 'async_trait>(
&'life0 self,
e: <Self::Endpoint as Endpoint>::Peer
) -> Pin<Box<dyn Future<Output = Option<Self::Endpoint>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Handles the message.

This future is dispatched on the executor directly, and does not block any other operations.

If the return value is Some, and the connection is still open when this method returns, the returned message is sent.

The message can be a single message or a request message. It is guaranteed to never be a response message, when called from transport::router.

Loading content...